Inherit from neuralintents.GenericAssistant to fix bugs with the package
